requirements.txt
absl-py==2.4.0
astunparse==1.6.3
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
Flask==3.0.3
flatbuffers==25.12.19
gast==0.7.0
google-auth==2.49.2
google-auth-oauthlib==1.3.1
google-pasta==0.2.0
grpcio==1.80.0
gunicorn==21.2.0
h5py==3.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==2.15.0
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.1.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.19.0
packaging==26.1
pandas==2.2.2
protobuf==4.25.9
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.20.0
PySastrawi==1.2.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
six==1.17.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-cpu==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
Werkzeug==3.1.8
wrapt==1.14.2
